requirements.txt
firebase-admin==6.9.0
google-api-core==2.25.1
google-api-python-client==2.120.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.49.1
httplib2==0.22.0
oauthlib==3.2.2
requests==2.32.3
requests-oauthlib==2.0.0
protobuf==4.22.0
rsa==4.9.1
cachetools==5.3.1
six==1.17.0
urllib3==2.3.0
Django==5.2.3
django-cors-headers==4.7.0
numpy==1.24.4

